+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Remove support for EC_P224 keys
 - Add support for EC_P256K1, BrainpoolP256, BrainpoolP384 and BrainpoolP512 keys
+- Remove `enable_set_attribute_value` config option
 
 ## [1.7.2][] (2025-07-25)